django_project_25/requirements.txt
Django~=3.2.15
python-dotenv==0.21.0
mysqlclient==2.1.1
Pillow==9.3.0
djangorestframework==3.14.0
drf-yasg==1.21.4
django_debug_toolbar==3.8.1
